% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_taxoname.R
\name{find_taxoname}
\alias{find_taxoname}
\title{Locate and extract taxonomic names from given input files}
\usage{
find_taxoname(filepath, filename, type, encoding = "unknown", output_name = "FALSE")
}
\arguments{
\item{filepath}{Required. The path of the file which the data is to be read from.
If it does not contain an absolute path, the file name is relative to the current
working directory.}

\item{filename}{Required. The name of the file which the data is to be read from.}

\item{type}{Required. Currently accept 'txt', 'docx', and 'pdf' format files.}

\item{encoding}{Optional. The encoding method of the input file. Default value is
'unknown'.}

\item{output_name}{Required. The path and name of the file for writing. If it does not
contain an absolute path, the file name is relative to the current working directory.}
}
\value{
A data frame containing the result of finding and reorganizing taxonomic names
in the input file into standard format.

A TXT file written from the above data frame and each line of this file contains one
entry of taxonomic names.
}
\description{
\code{find_taxoname} locates and extracts taxonomic names from txt, docx, pdf or html
files and reorganizes the taxonomy names into standard order: genus, species, subspecies,
author&year, distribution. The function can output the result to a txt file and each row
of the file is one entry of a taxonomic name. The result txt file of this function can be
further processed into a tabular format in csv which contains more detailed information
using function \code{parse_taxolist}.
}
\examples{
df <- find_taxoname(filepath = "D:/code/Parser/Parser_for_Biodiversity_Checklists/Examples/input_data",
                    filename = "taxo01.txt",
                    type = "txt",
                    output_name = "D:/code/Parser/Parser_for_Biodiversity_Checklists/Examples/output_data/taxo01_output")

}
